Create a class library for consumers to implement manifest-driven code signing
Currently, code signing is done by folder location and convention. This is fragile and cause tight coupling
between the code building outputs and the signing tools. This change introduces a set of abstractions
for working with the signrequest.xml file. Sign tool implementers can read this file and use it
to drive code signing implementations.