Implement MicroBuild code signing
This transforms MSBuild items (see docs/Signing.md) into a manifest which is then invoked by the Roslyn sign tool. In environments with MicroBuild enabled, this will run code signing. In local developer builds, this tool will do a dry run that validates the assemblies in .nupkgs are account file in the sign data